Fix blocker spawned by dead persons
